feat: [윤희지] app bar, drawer 게시판 이름 통일 [HOME-ALONE26]
