+++ p4th/HeejiYun/frontend/lib/pages/boards/board_list_page_recipe.dart
@@ -1,23 +1,23 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
-import 'package:frontend/components/forms/BoardListView.dart';
+import 'package:frontend/components/forms/board_list_view.dart';
 import 'package:frontend/components/custom_app_bar.dart';
 import 'package:frontend/components/custom_drawer.dart';
 import 'package:provider/provider.dart';
 
-import '../utility/board_list_provider.dart';
+import '../../utility/board_list_provider.dart';
 
-class BoardListPageAsk extends StatelessWidget {
-  BoardListPageAsk({Key? key}) : super(key: key);
+class BoardListPageRecipe extends StatelessWidget {
+  BoardListPageRecipe({Key? key}) : super(key: key);
 
   late BoardListProvider _providerTest;
 
   @override
   Widget build(BuildContext context) {
     _providerTest = Provider.of<BoardListProvider>(context, listen: false);
-    _providerTest.loadAskBoards();
+    _providerTest.loadRecipeBoards();
     return Scaffold(
-        appBar: CommonAppBar(title: "자취인 질문 게시판"),
+        appBar: CommonAppBar(title: "1인분 게시판"),
         drawer: CustomDrawer(isLogin: true,),
         body: Consumer<BoardListProvider>(
           builder: (context, provider, widget){
